requirements.txt
# The-Machine: Core and Extension Dependencies
#
# ⚠️ We STRONGLY recommend using Anaconda/conda to manage your environment, especially for PyTorch and GPU support.
# See the README for full conda setup instructions.
#
# If you must use pip, install the correct torch/torchaudio version for your system from https://pytorch.org/get-started/locally/
#
# This requirements.txt is for all core and extension features. Some extensions may require additional packages.

# Core dependencies
# (Install torch/torchaudio via conda for best compatibility)
torch>=2.0.0
# For GPU, install torch/torchaudio with CUDA support from https://pytorch.org/get-started/locally/
torchaudio>=2.0.0
tqdm
rich
numpy<=2.1.*  # Pinned: Numba/NeMo require ≤2.1 (NumPy 2.2+ breaks RNNT numba jit)
soundfile
mutagen
pydub
scipy
colorama
pyyaml
requests

# Hugging Face ecosystem
transformers>=4.30.0
datasets
huggingface_hub

# CLAP (audio event detection, for extensions only)
# (CLAP is available via transformers >=4.30.0)

# Parakeet ASR (NVIDIA NeMo)
# NVIDIA NeMo – ASR + TTS support
nemo_toolkit[asr,tts]>=1.22.0

# Whisper ASR
openai-whisper

# Audio separation
# (audio-separator is a third-party tool, install as needed)

# For YouTube/URL audio ingestion
yt_dlp

# If you use Jupyter or notebooks
ipython

# Optional: for logging, YAML config, etc.
# pyyaml, colorama already included 

# FlashSR super-resolution
git+https://github.com/jakeoneijk/FlashSR_Inference

# Better Profanity
better_profanity

# Added for Auto-Editor extension wrapper
auto-editor>=28.0.0

# Vision-language & image processing
pillow>=10.0.0

# Faster Whisper
faster-whisper>=0.10.0

# ---
# Conda installation instructions (recommended for PyTorch):
#
# 1. Create a new environment:
#    conda create -n themachine python=3.10
#    conda activate themachine
#
# 2. Install PyTorch (choose the right CUDA version for your system):
#    See https://pytorch.org/get-started/locally/ for the latest command.
#    Example (CPU only):
#      conda install pytorch torchaudio cpuonly -c pytorch
#    Example (CUDA 11.8):
#      conda install pytorch torchaudio pytorch-cuda=11.8 -c pytorch -c nvidia
#
# 3. Install the rest of the dependencies:
#    pip install -r requirements.txt 